+++ tests/testthat/test-text_primitive.R
@@ -64,3 +64,40 @@ test_that("invalid operations should error", {
     expect_error(as.environment(x),
                  "'as.environment' is invalid for text objects")
 })
+
+test_that("setting length on invalid text should fail", {
+    x <- letters
+    expect_error(`length<-.corpus_text`(x, 5), "invalid text object")
+})
+
+test_that("setting invalid length should fail", {
+    x <- as_corpus_text(letters)
+    expect_error(length(x) <- NULL, "'length' cannot be NULL")
+    expect_error(length(x) <- "1", "'length' must be numeric")
+    expect_error(length(x) <- c(1, 1), "'length' must have length 1")
+    expect_error(length(x) <- NA, "'length' cannot be NA")
+    expect_error(length(x) <- NaN, "'length' cannot be NaN")
+    expect_error(length(x) <- -1, "'length' cannot be negative")
+    expect_error(length(x) <- 2^53 + 2, "'length' cannot be above 2\\^53")
+})
+
+
+test_that("setting short length should work", {
+    x <- as_corpus_text(letters)
+    length(x) <- 10
+    expect_equal(x, as_corpus_text(letters[1:10]))
+})
+
+
+test_that("setting same length should work", {
+    x <- as_corpus_text(letters)
+    length(x) <- 26
+    expect_equal(x, as_corpus_text(letters))
+})
+
+
+test_that("setting long length should work", {
+    x <- as_corpus_text(letters)
+    length(x) <- 30
+    expect_equal(x, as_corpus_text(c(letters, rep(NA, 4))))
+})